# Speed up initial minimal analysis by ≈25%