Speed up initial scan by exiting from the loop early if the line contains more than 3 alphanumerical characters.

<details><summary>Measurements</summary>
<p>

Codebase: [SerenityOS](https://github.com/SerenityOS/serenity)

Hardware: MacBook Pro 14" (M1 Pro)

## With the change
### Run 1
```
2023-10-07 19:54:33,337 Creating server...
2023-10-07 19:54:33,338 Starting worker thread...
2023-10-07 19:54:33,340 Serving on http://0.0.0.0:53385
2023-10-07 19:54:33,753 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 19:54:46,419 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 19:54:46,419 Note, 835884 chunks need to be analyzed for optimum performance.
```
13.082 sec before minimum number of chunks is analyzed

### Run 2
```
2023-10-07 20:02:01,040 Creating server...
2023-10-07 20:02:01,041 Starting worker thread...
2023-10-07 20:02:01,043 Serving on http://0.0.0.0:53459
2023-10-07 20:02:01,555 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 20:02:13,640 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 20:02:13,640 Note, 835884 chunks need to be analyzed for optimum performance.
```
12.6 sec before minimum number of chunks is analyzed

### Run 3
```
2023-10-07 20:02:37,084 Creating server...
2023-10-07 20:02:37,085 Starting worker thread...
2023-10-07 20:02:37,088 Serving on http://0.0.0.0:53480
2023-10-07 20:02:37,607 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 20:02:49,111 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 20:02:49,111 Note, 835884 chunks need to be analyzed for optimum performance.
```
12.027 sec before minimum number of chunks is analyzed

Average 12.57 sec

## Without

### Run 1
```
2023-10-07 19:59:01,155 Creating server...
2023-10-07 19:59:01,156 Starting worker thread...
2023-10-07 19:59:01,159 Serving on http://0.0.0.0:53418
2023-10-07 19:59:01,718 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 19:59:17,896 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 19:59:17,896 Note, 835884 chunks need to be analyzed for optimum performance.
```
16.741 sec before minimum number of chunks is analyzed

### Run 2
```
2023-10-07 19:59:56,583 Creating server...
2023-10-07 19:59:56,584 Starting worker thread...
2023-10-07 19:59:56,586 Serving on http://0.0.0.0:53427
2023-10-07 19:59:57,099 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 20:00:13,248 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 20:00:13,248 Note, 835880 chunks need to be analyzed for optimum performance.
```
16.665 sec before minimum number of chunks is analyzed

### Run 3
```
2023-10-07 20:31:23,370 Creating server...
2023-10-07 20:31:23,371 Starting worker thread...
2023-10-07 20:31:23,373 Serving on http://0.0.0.0:53690
2023-10-07 20:31:23,900 Checking repository for new changes
Analyzing source code: 0it [00:00, ?it/s]
2023-10-07 20:31:40,124 Analyzed the minimum number of chunks needed to operate. 
2023-10-07 20:31:40,124 Note, 835884 chunks need to be analyzed for optimum performance.
```
16.754 sec before minimum number of chunks is analyzed

Average 16.72 sec

Total improvement: 1 - 12.57/16.72 ≈ 0.25

</p>
</details>